MODULE 4
Trade Call Prediction using Classification
In this module, we'd be covering the concept of classification and utilize our skills to solve the following queries – (Stock Price =
Close Price)
Problem Statements
4.1 Import the csv file of the stock which contained the Bollinger columns as well.
Create a new column 'Call' , whose entries are -
'Buy' if the stock price is below the lower Bollinger band
'Hold Buy/ Liquidate Short' if the stock price is between the lower and middle Bollinger band
'Hold Short/ Liquidate Buy' if the stock price is between the middle and upper Bollinger band
'Short' if the stock price is above the upper Bollinger band

Now train a classification model with the 3 bollinger columns and the stock price as inputs and 'Calls' as output. Check
the accuracy on a test set. (There are many classifier models to choose from, try each one out and compare the accuracy
for each)
